refacto : ajout de isAdmin aux profs

+++ app/routes/teachers_routes.py
@@ -1,9 +1,9 @@
-from flask import Blueprint, request, redirect, url_for
+from flask import Blueprint, request, redirect, url_for, flash
 from werkzeug.security import generate_password_hash
-from flask_login import login_required, current_user
-
+from flask_login import login_required
 from app.models import Teacher
 from app.repositories import TeacherRepository
+from app.services.singletons.logger_singleton import Logger
 
 teachers_bp = Blueprint("teachers", __name__)
 
@@ -16,7 +16,11 @@ def add_teacher():
     password = request.form.get("password")
 
     if not all([first, last, email, password]):
-        print("Tous les champs sont requis.", "error")
+        flash("Tous les champs sont requis.", "danger")
+        return redirect(url_for("main.studio"))
+
+    if TeacherRepository.get_by_email(email):
+        flash("Un professeur avec cet email existe déjà.", "warning")
         return redirect(url_for("main.studio"))
 
     teacher = Teacher(
@@ -24,24 +28,34 @@ def add_teacher():
         last_name=last,
         email=email,
         password_hash=generate_password_hash(password),
-        role="prof"
+        is_admin=False
     )
     TeacherRepository.create(teacher)
-    print("Professeur ajouté avec succès !", "success")
+    Logger().log(f"Professeur {teacher.email} ajouté avec succès !")
     return redirect(url_for("main.studio"))
 
 @teachers_bp.route("/studio/teacher/delete/<int:id>")
 @login_required
 def delete_teacher(id):
     teacher = TeacherRepository.get_by_id(id)
-    TeacherRepository.delete(teacher)
-    print("Professeur supprimé.", "success")
+    if teacher:
+        Logger().log(f"Professeur {teacher.email} supprimé.")
+        TeacherRepository.delete(teacher)
     return redirect(url_for("main.studio"))
 
 @teachers_bp.route("/studio/teacher/edit/<int:id>", methods=["POST"])
 @login_required
 def edit_teacher(id):
     teacher = TeacherRepository.get_by_id(id)
-    TeacherRepository.edit_teacher(teacher, request.form.get("first_name"), request.form.get("last_name"), request.form.get("email"))
-    print("Professeur modifié avec succès.", "success")
+    if not teacher:
+        flash("Professeur introuvable.", "danger")
+        return redirect(url_for("main.studio"))
+
+    TeacherRepository.edit_teacher(
+        teacher,
+        first_name=request.form.get("first_name"),
+        last_name=request.form.get("last_name"),
+        email=request.form.get("email")
+    )
+    Logger().log(f"Professeur {teacher.email} modifié avec succès.", "success")
     return redirect(url_for("main.studio"))
